requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pillow==11.2.1
Faker==37.3.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.5.0
python-keycloak==5.3.1
PyJWT==2.9.0
requests==2.31.0
django-keycloak-auth==1.0.0
python-jose==3.4.0
#djangorestframework-keycloak

redis
jwt
django-redis
keycloak